Fixed two rare pathfinding bugs.
The non-uniform path code had a two minor mistakes.

Firstly, the possibility of a non-uniform cost was being ignored on the
first step of each turn (apparently I was intending to turn that into a
special case, for efficiency, but I never got around to it).

Secondly, when forward-propagating changes on a path, the path cost was
being recalculated before the parent node was updated, and so in the
path was inaccurate when tracing backwards to calculate the non-uniform
cost of the path.

These two bug only affected a sub-set of the cases that were already
affected by the original bug that the non-uniform stuff was meant to fix
anyway - so they were really very rare.