Changed the danger interrupt condition for workers.
Originally, workers would stop if there was any danger within two steps.
Awhile back I changed this so that danger was ignored if the worker's
command was issued while the worker was already in danger. But recently
I've found that this produces some unintuitive outcomes which may cause
players to have their workers captured. I've now changed the condition
so that if the command is issued while in danger, the worker will still
be interrupted if there is danger within 1 step - but not 2 steps.

If this new version is later found to not be satisfactory, I guess I'll
have to change it again so that there are three different modes for the
workers - interrupt for 2 steps, for 1 step, or not at all - depending
on ... stuff.