Corrected a mistake in the final spread condition of AI_bestReligion.
The calculation of the extent that the religion has spread was using the
maximum of the target cities and the player's cities, which meant that
players with fewer cities than the target may be unable to reach the
spread requirements even if the religion was in all of their cities.

I've changed it to use the _minimum_, of
(1.5*target cities + 1, current cities + 1).