Byzantine hippodromes now allow two artists.
Before this change, the hippodrome was the only unique building which
had a disadvantage over the base building. I've changed it to allow two
artists so that it matches theatres and other theatre replacements.

(This change removes some of the flavour of the building, but I think it
was a bit unfair that the byzantine couldn't get these artists.)