Added some plot culture dependence to CvCity::getLiberationPlayer.
Originally the liberation player depended almost entirely on how far the
city is from the capitals of other players. I've tweaked this so that it
has some dependence on the plot culture as well - but the distance from
the capital is still the dominant component.

I've also replaced some large if blocks with if! continue conditions, to
flatten the code so that it is more readable (as is consistent with the
style I've used elsewhere).