Fixed a minor bug in AI_doWar relating to automatic peace deals.
The bug was that when making a peace deal, 'break' was used instead of
'continue', which meant that the AI would ignore peace conditions with
other teams after making peace with any one team.

(This was a very long-standing bug from the original BtS code and bbai
code. The bug was made more obviously visible by my style adjustments.)