Players can now tell their AI allies start war preparations.
Human players can now tell their AI team members and AI vassals to start
preparing for war against a particular team. When used, the AI will
adopt WARPLAN_PREPARING_TOTAL, and when war is eventually declared this
will be changed to WARPLAN_TOTAL. There is also a diplomacy option to
tell these AI allies to cancel the war preparations.

(Note: if this feature is not used the AI will use WARPLAN_DOGPILE, as
usual. The AI cannot currently switch between different warplans once
war is declared.)

These changes includes new diplomacy options in the "let's discuss
something else screen", new text for these diplomacy options and AI
responses, and a new type of 'diplomacy event' to be handled by the DLL.

(Note: The new diplomacy text was translated by a computer, and I
haven't verified that it's correct or appropriate.)