Blocked the AI from considering very early wars.
The AI will no longer consider starting a new war if less than 30 turns
have past (scaled by game speed based on 'ibarbPercent'), and if there
are fewer cities globally than half of the total cities target. ie.
Number of civs * target number of cities / 2. (Generally, I expect that
the turns limit will expire before the cities target is reached - except
perhaps in advanced start games, or in games where players start with
additional settlers or something like that.)

Note that in the original BtS AI there was a convoluted set of
conditions, based around AI_STRATEGY_GET_BETTER_UNITS, which prevented
the AI from starting early wars. Quite a long time ago I changed the
conditions of that AI strategy, and one of the side effects of my change
was that it no longer prevented early wars. That was an unintended side
effect, and so I'm addressing it now.