Made some structural changes to CvPlayerAI::AI_baseBonusVal
The motivation for these changes was that I wanted to make the AI's
resource trading a bit smarter. In particular, I wanted them to not
value strategic resources as highly if they already have another
resource to fulfil the same requirements. (eg. don't value copper as
highly if they already have iron.)

Unfortunately, after writing the changes I realised that they were
likely to cause OOS problems. So I've commented out the parts of the
code that I think could cause problems - and this means there is very
little functional difference from the original code. The only reason I'm
keeping the new code at all is that I might want to use it in the future
if I ever change the cache system such that OOS is no longer a problem.