Rewrote most of AI_assaultSeaTransport.
The AI will now be able to consider amphibious attack landings on
non-city plots. Previously the AI would never do this, and the fact that
it would never do it was easy for players to exploit.

The AI now uses a full strength comparison for evaluating their
amphibious attacks, rather than just counting units.

Added a generalised unit hash function, CvUnitAI::AI_unitBirthmarkHash,
for when we need a hash that depends on something other than a plot.
(ie. when we don't want to use AI_unitPlotHash.)