Added randomness to the AIs decision to 'cherry pick' target cities.
When not in 'total war', the AI is a bit more selective in its choice of
cities to attack. I think this works well, except that in some
circumstances it is possible for players to use this to manipulate which
city the AI targets. I've added some randomness so that the AI sometimes
disables its cherry-picking of cities so that players can't manipulate
the AI indefinitely.