Fixed the AutoPlay module disabling event handlers from other mods.
I haven't looked much at the code for the AIAutoPlay module, so
hopefully I haven't misunderstood how it is suppose to work...

It looks like AIAutoPlay overrides the event handlers for new turns and
for new cities so that popups don't appear while autoplay is running.
I've edited the overriding events so that they basically just call the
original event handlers if autoplay is not active.

It would be better to block the popups directly rather than blocking the
event handlers. The best way to do this would be to add autoplay checks
to 'onCityBuilt' in CvEventManager.py, and to 'turnChecker' in
CvTopCivs.py. But I'm not going to make those changes right now, because
I'd rather not have to add those two files to this mod for such a
trivial reason. I might do it later if there are still problems.