Added additional checks to CvGame::calculateSyncChecksum
Also, the CvMainInterface.py now shows a number to indicate which
checksum has failed based on the current timeslice. (Unfortunately, I
can't guarantee that _all_ the failed checks will be listed, but at
least we'll know that the ones that are listed have failed.)

These changes are only for debugging.