Citizens are once again unassigned before automatic reassigning.
CvCityAI::AI_assignWorkingPlots now removes all citizens from their
plots before reassigning them all one by one. This was the original
behaviour, but I had disabled it for efficiency reasons. I've reverted
the change because my testing has shown that it sometimes produces
slightly better citizen assignment, and it doesn't have a huge impact on
speed. (I had hoped that my rewrite of the juggleCitizen method would be
good enough to not require the initial unassigning, but unfortunately
that's not the case.)