The open list in KmodPathFinder is now a vector, to improve speed.
Although it seems natural to use a linked list for this, I recently
found out that linked lists are generally slower than vectors even when
deleting and inserting in the middle of the list. I did a quick
experiment to test it, and sure enough, the vector was faster.

Thanks to Bjarne Stroustrup for the tip.