Restructured pathfinder usage to not use getPathLastNode.
FAStarNode pointers returned by the pathfinder become invalid when the
pathfinder is destroyed. Because of this, I consider it bad practice to
use these pointers. In fact, the the pathfinder shouldn't allow any kind
of external access to them at all. Not only could misuse result in
segfaults, but the pointers could also be misused in a way to mess up
the internal state of the pathfinder...

So, in this patch, almost all usage of FAStarNode has been removed from
outside of the pathfinder. This will have no effect on the execution of
the game, but it may help to reduce bugs in the future.

I've disabled CvUnit::getPathLastNode and
CvSelectionGroup::getPathLastNode, and replaced their usage with new
pathfinder methods such as KmodPathFinder::GetFinalMoves and
KmodPathFinder::GetPathTurns.

There are only a couple of places where FAStarNode is still used. One is
in CvUnitAI::AI_solveBlockageProblem, which doesn't use the
KmodPathFinder, so I probably won't change it. The other is
CvUnitAI::AI_considerPathDOW. To correct this, I intend to create a new
method to give access to the completed path without exposing the
internal nodes. (But I haven't done that yet.)

I've also removed the KMOD_PATH_FINDER define and preprocessor
conditions. The original bts pathfinding code has been deleted.