Fixed units getting bumped when simultaneously attacking and defending.
This fixes a bug which sometimes caused units to move onto plots which
still had enemy defenders.

The root of the bug was that when two units attempt to attack each other
at the same time, the game failed to track which unit was in the process
of attacking.

I've corrected this by clearing attackPlot when a unit is defending. And
I've added some additional conditions and asserts to prevent similar
problems in the future.