Fixed some problems with cultural cities in the victory screen.
The cities are now grouped be team rather than by player, to match the
actual victory conditions. (A team needs 3 legendary cities to win. It
doesn't matter if the cities aren't owned by the same player.)

This means that allied cities can now be shown amongst your team's best
cities, and they can not be considered as rival cities.