Further adjustments to food handling for citizen work assignment.
Rewrote the 'iStarvingAllowance' calculation in AI_yieldValue. The
previous implementation only triggered for AI players with 'avoid
growth' turned on - which was impossible. The new implementation allows
some food loss when the bar is above 80% full, both for AI and human
players (with slightly different rules).

Simplified and tweaked some of the other starvation calculations.

AI_juggleCitizens now uses "bRemove=true" for considering both worked
and unworked plots. This is a kludge. It makes more sense for bRemove to
be false when considering unworked plots. But the fact is that however
we do it is going to be somewhat unhelpful, because what we really need
is a direct food comparison between plots. I think the current code is
good enough for now, but if I become dissatisfied with it again, I need
to remember that a direct food comparison is the only way it is going to
get better!