Added simple direct food comparison to AI_juggleCitizens.
Starvation cost is now calculated separately from the rest of the job
value, so that it can be taken into account via a direct comparison
between the old and new jobs before switching.

Note: one problem with this system is that a high-value plot which would
cause starvation can prevent the consideration of a lower value plot
which might not cause starvation. I'm intend to rewrite the function to
fix this problem.