Fixed OOS bugs in some random events.
There were a few events which involved sorting lists that had memory
dependant values in them. eg. list of (distance, city_pointer) pairs.
This meant that in some situations memory addresses were compared, which
sometimes resulted in OOS (because the memory address won't be the same
on all computers, thus the lists might have different orders).

I've changed it so that only the first member of the tuple is used to
sort the list. The memory address will be ignored, and so that should
solve the problem.