Fixed a bug in the evaluation of food difference in AI_yieldValue.
When AI_yieldValue was used only for plots, I chose to not count growth
value for plots which did not offer any yield than citizens would
consume (eg. a grassland plot with no river and no improvements).

AI_yieldValue is now used for direct comparisons between plots, and so
we can no longer assume that a yield of only 2 food is useless - because
it might be a _difference_ of two food.

Note: the best way to restore the behaviour of ignore growth value on
bad plots would be to alter iGrowthValue in AI_plotValue.