Increased tendency for settler and missionary boats to wait in cities.
Units are more likely to load into boats if the boats are waiting in a
city - and so to help the AI spread corporations, I've added a new
condition for using missionary boats for exploring.

The boats will now only explore if there aren't alrady enough explorers
in the area.