Fixed some minor problems in AI_load.
BBAI added some code to AI_load which checked for enemy territory in the
range of the boat to be loaded onto. I've added conditions to this check
so that it doesn't apply to missionary boats or settlers boats; and I've
made it search from the boat's position rather than the unit to be
loaded's position.

(It's still a pretty rough and inefficient system, but it's probably
good enough for now.)