Adjusted how success probability is shown in the space-ship screen.
Success probability is now shown next to the launch button, rather than
in the opposite corner of the screen. Also, the probability is now
displayed even when adding new spaceship parts; whereas previously it
was hidden.

(I think it's bad that the original version allowed the player to launch
their ship without ever seeing the success probability - due to the
probability being hidden when parts are being added. That's what's
motivated these changes.)