Added gold slider estimate for use in cultural victory strategies.
Added method CvPlayerAI::AI_estimateBreakEvenGoldPercent.
AI_estimateBreakEvenGoldPercent estimates the percentage needed on the
gold slider to pay for all on-going expenses.

AI_estimateBreakEvenGoldPercent is now used for
AI_calculateCultureVictoryStage, and AI_updateCommerceWeights to provide
a more stable estimate of how many turns it would take to achieve a
cultural victory.

(I'm tempted to use AI_estimateBreakEvenGoldPercent in
AI_commerceWeight, to increase the value of gold when the civ's economy
is weak. But I'm worried that it will be a bit too slow; and that it
will be double-counting with the AI's "special yield" modifiers.)