Capitulation power is now determined after cities are reassigned.
Vassals are able to free themselves if their land drops below some
percentage of what it was when the capitulation deal was signed.

That original land size is now determined at the end of deals rather
than at the start, so that capitulation deals which involve city
transfers don't result in vassals being able to free themselves
immediately.

(Also made some style adjustments.)