Rewrote AI_techValue and helper functions to establish a common scale.
Changed all the numbers in AI_techValue such that 4 ~= 1 commerce per
turn. Tech values should now scale with civ size.

Adjusted AI_techUnitValue, AI_techBuildingValue, and
AI_obsoleteBuildingPenalty to match the new scale.

Also made some actual (non-scale) adjustments and bug fixes to
AI_techValue. (For example, the number of existing improvements was not
counted correctly in the previous code.)

Note: there are a lot of changes here; and most of it deals with
arbitrary numbers. It's quite likely that there will need to be further
fixes or adjustments to this later on.