Some adjustments to the new AI_techValue numbers.
Greatly reduced the random component of AI_techValue (it looks like I'd
forgotten to rescale it.)

AI_techBuildingValue now reduces values to 2/3 if the player only had 1
city, and is still on the first era.

Upgradable improvements now use average of first and last yield values
for the AI_techValue evaluation.