Inflation rate now increases with global tech state.
Added function CvPlayer::updateInflationRate, and replaced
calculateInflationRate with getInflationRate. (Mods which used
calculateInflationRate in their python code will need to be updated to
use getInflationRate instead.)

Inflation is now cached for each player. (The new inflation calculation
is significantly slower than the old.) Note: saves from before this
change will have zero inflation for the first turn, due to limitations
of the load process.

The effects of inflation increase with each turn as before, but now the
'effective turn' number used in that calculation depends on how many
technologies have been discovered globally.

The goal of this change is to increase inflation in games with a high
tech-rate, so that the inflation rate is suitably high when civs have
the advantage of late-game buildings and plot improvements.