Reduced chance of spies being caught while exploring friendly land.
The "recent mission" term in the spy intercept probability is applied if
the spy has not been fortified or if there is more than one friendly spy
on the same plot. This meant that spies exploring in friendly (foreign)
territory often got caught.

I've changed the 'recent mission' penalty so that it now only applies if
one of the following conditions are met:
 * More than one spy is on the plot.
 * The players are at war.
 * A counter-espionage mission is in effect.
 * A counter-spy (or security bureau) is on the same plot.

Roughly speaking, this means that probability of catching friendly spies
is a lot lower, but the probability of catching hostile spies is
unchanged.