Made several adjustments to AI_techValue.
Greatly increased the scale of AI_techBuildingValue (it is now closer in
impact to how it was before techValues were rescaled).

Increased the value of extra trade routes, by including the average
commerce multiplier.

Decreased the base value of founding a religion - but partially offset
this decrease by increasing the value when there are no nearby
religions.

Slightly decreased the value of enabling wonders.

Fixed the value calculation for free tech so that it used the iRoll
value rather than a flat number.