Implemented "SpecialistExtraCommerce" xml field for technologies.
Techs can now give a civ-wide commerce bonus to all specialists in the
same way that civics can. The new xml field, SpecialistExtraCommerces,
is optional - so techs which don't use this new field do not need to be
updated.

The AI has been taught to evaluate the new bonus. And the python tech
screen has been updated to show an icon and mouse-over text for techs
with the new bonus.

I've also changed the tech boxes in the research screen to be slightly
wider, so that the box for Scientific Method will be wide enough in
include one more bonus.