Fixed the initial selection on the graphs screen pull-down menu.
The graph screen pull-down menu now correctly shows the currently
selected graph when the screen is opened.

Previously, the menu always showed "Score" when the screen is freshly
opened, even though the actual graph being displayed might be something
else. (The graph being displayed was stored from the previous time the
screen was open, but the selection on the menu always reverted to
"score".)