Adjusted cultural revolt probability and properties.
Increased the base change of revolt from 10% to 15%. Decreased the
duration of revolts from 3+2*(previous revolts) to 2+(previous revolts).
Increased the number of 'warning revolts' from 1 to 2. ie. Cities will
now flip on the 3rd revolt rather than the 2nd.

The net effect of these changes is that the average number of turns to
flip a city will be unchanged, and the number of turns spent in revolt
will be similar to before - but the chance of each revolt is a bit
higher and so it's a bit more predictable and less dependant on luck.