Tided the xml used in CvGameTextMgr::setUnitHelp
CvGameTextMgr::setUnitHelp was previously using "<img=%S size=16></img>"
for all the promotion icons. I've changed it to "<img=%S size=16 />"
instead, which is a bit shorter and cleaner.

Realistically it shouldn't make any difference; but I've encountered a
rare bug which messes up the on-screen text from setUnitHelp. I've spent
quite a lot of time debugging it, and my best guess is that it is a bug
in the xml processor (which I assume is in the .exe, because I can't
find it in the dll code).

In the save game I was using to reproduce this bug, changing the img
tags to the shorter version removed the problem. I don't think the bug
is actually related to that though. I think it's something to do with
the precise length of the text or something like that - because changing
pretty much anything at all removed the problem. I'm just committing
this change as a kind of memento of the time I've spent debugging this
problem.