Relaxed conditions in AI_retreatToCity, to fix AI indecision.
Due to special conditions applied to the unit's current plot, the AI
still preferred to move to a new city rather than stay in the current
one. Those special conditions have been removed, so that the units
retreating to a city will generally be happy to just stay where they
are.

(I've corrected similar problems in AI_retreatToCity in the past; but
this change is actually the most straight forward.)