Fixed several problems with slavery evaluation in AI_yieldValue
There were some mistakes which mean slavery was sometimes counted when
it shouldn't be, and sometimes wasn't counted when it should be; but the
biggest problem was that it was a factor of 100 too small... and so it
was essentially being ignored.

I've fixed the scale problem and adjusted the conditions.