Fixed misuse of maxMoves() when intending baseMoves()
There were several places where CvUnit::maxMoves() was used when
baseMoves() was intended. This was mostly in old naval AI code. The main
effect of this was to increase the search time for the AI by searching
around 400 times more plots than intended.

baseMoves() is the number of tiles a unit can move. maxMoves() the total
"movement points", which is baseMoves() * a constant.