Fixed potential pointer invalidation in CvSelectionGroup::groupMove
The selection group can be messed up when units move. eg. If a city is
captured, the change in culture could bump the remaining units.

To get around this, CvSelectionGroup::groupMove uses a copied list of
units rather than the live list of units.

Thanks to alberts2 and DarkLunaPhantom for the tip.