Rewrote CvPlayerAI::AI_counterPropose to get more balanced deals.
AI_counterPropose now worked by evaluating all possible trade items
before adding them to the trade, so that items can be added in the order
most likely to result in a balanced deal.

In addition to this, the AI will now be slightly more generous to humans
in the sense that the automatic counter proposals from human players
will no longer add gold if the AI would accept the deal without gold.
These automatic counter proposals are the only time the AI will allow a
discounted trade with a human. This change means that the AI may offer a
trade deal which favours a human player even if the human player has
enough gold to balance the deal. Previously, if the human player had a
lot of gold on hand, the AI would never offer a discounted trade. (Also
note that in the original BtS code, the AI would never offer a
favourable deal to a human under any circumstances.)

This was a significant rewrite, so there might be some bugs. However, my
main concern is that the new version might actually work 'too well' in
the sense that AI players might make significantly more trade deals with
each other, and thus be stronger than before. We'll see if further
changes are needed.