Adjusted AI_counterPropose to prefer an unfair deal over no deal.
When considering the iValueForThem > iValueForUs case, we were trying to
get as close to a balanced trade as possible. However, this meant that
when the fairest deal would favour a human 'them', the AI would decline
to offer any deal at all.

I've altered the behaviour so that the top priority is to favour 'us',
and getting a fair deal is secondary; rather than the other way around.
This means fair AI-AI deals might be less likely; but counter-proposals
to human trade offers will work better. And that's the main aim.