Replaced several misuses of CvUnit::getMoves with baseMoves
There were a few places in the AI which used getMoves when they really
wanted baseMoves. The main side effect of this was that many searches
too longer than they should; but it also resulted faulty reasoning for
some decisions (eg. thinking the enemy was in range when it wasn't).

It's very frustrating for me that CvUnit::getMoves doesn't mean the same
thing as CvUnitInfo::getMoves. I'd rename one of them, but they are both
quite engrained.