Many changes to AI_techValue.
Disabled the random value bonuses for techs with iPathLength > 1. (Raw
bonus, and bonus from wonders). ie. techs for which not all prereqs have
been met no longer get these random bonuses.

Note: This is partially to help the AI focus on real benefits; but a big
reason for removing this randomness is a technical one. I'll soon to
changing the way multi-step research paths are evaluated. Having
randomness in prereqs could upset the evaluations because the values
will be used for several techs.

Decreased random wonder value.

Increased the value of gold trading.

Changed several constant values to scale with the number of cities.
(These things don't necessarily have anything to do with the number of
cities. We're only multiplying by cities so that the value scales like
things which are evaluated properly.)

Increased chop value.

Moved project evaluations to a new function: AI_techProjectValue.

Increased value of civic improvements.

Decreased the value of religions.

Rewrote evaluatation of bonus techs. It's now based on the average of
the max tech cost and average tech cost for currently researchable
techs.

Renamed bIgnoreCost to bFreeTech, to better reflect what the argument is
used for. Rescaled the return value so that it is roughly comparable for
free and non-free tech.

Renamed "tech whoring" to "tech ground-breaking", to better reflect the
concept. (ie. getting the tech first, for better trade options.)
Slightly increased the value.