Restored WHEOOHRN scoreboard indicator.
Note: if you have existing K-mod config files, then you'll need to add
"M" somewhere to the advanced scoreboard format string to see this
indicator. I recommend putting it next to the W.

The WHEOOH indicator was removed before the first version of this mod
was released; with the reasoning that the AI should not broadcast its
war-plans to the player in that way.

It is still possible to get WHEOOH info from the AI, but they will only
give that response under certain diplomatic conditions. With that in
mind, I've restored functionality of the indicator with the same
conditions that are on the diplomacy screen. This is something that a
lot of people have asked for.