Fixed two bugs in new tech AI (thanks to f1rpo)
There was a typo which prevented prereqs from being added to the tech
path correctly. (This could sometimes result in mis-evaluations of tech
paths, but usually would be unnoticeable.)

Secondly, the prereq check previously would stop when the path was
full. This could cause the check to fail if two techs used the same
prereq. This is fixed by not stopping until the path is over-full.